Troca do padrão stateful vs stateless
