+++ app/src/main/java/br/com/alura/helloapp/ui/details/DetalhesContatoViewlModel.kt
@@ -1,8 +1,10 @@
 package br.com.alura.helloapp.ui.details
 
+import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import br.com.alura.helloapp.database.ContatoDao
+import br.com.alura.helloapp.util.ID_CONTATO
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -10,10 +12,13 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-class DetalhesContatoViewlModel(
+@HiltViewModel
+class DetalhesContatoViewlModel @Inject constructor(
     private val contatoDao: ContatoDao,
-    private val idContato: Long
+    savedStateHandle: SavedStateHandle
 ) : ViewModel() {
+
+    private val idContato = savedStateHandle.get<Long>(ID_CONTATO)
 
     private val _uiState = MutableStateFlow(DetalhesContatoUiState())
     val uiState: StateFlow<DetalhesContatoUiState>
@@ -26,18 +31,20 @@ class DetalhesContatoViewlModel(
     }
 
     suspend fun carregaContato() {
-        contatoDao.buscaPorId(idContato)?.let { contatoEncontrado ->
-            with(contatoEncontrado) {
-                _uiState.value = _uiState.value.copy(
-                    id = id,
-                    nome = nome,
-                    sobrenome = sobrenome,
-                    telefone = telefone,
-                    fotoPerfil = fotoPerfil,
-                    aniversario = aniversario
-                )
-            }
+        idContato?.let {
+            contatoDao.buscaPorId(idContato)?.let { contatoEncontrado ->
+                with(contatoEncontrado) {
+                    _uiState.value = _uiState.value.copy(
+                        id = id,
+                        nome = nome,
+                        sobrenome = sobrenome,
+                        telefone = telefone,
+                        fotoPerfil = fotoPerfil,
+                        aniversario = aniversario
+                    )
+                }
 
+            }
         }
     }
