+++ app/src/main/java/br/com/alura/helloapp/util/Constantes.kt
@@ -2,3 +2,5 @@ package br.com.alura.helloapp.util
 
 const val FORMATO_DATA_DIA_MES_ANO = "d/M/yyyy"
 const val FORMATO_DATA_EXIBIR = "dd/MM/yyyy"
+
+const val ID_CONTATO = "id_contato"
