+++ build.gradle
@@ -1,4 +1,3 @@
-
 plugins {
     id 'com.android.application' version '7.3.0' apply false
     id 'com.android.library' version '7.3.0' apply false
